Merge values from Forwarded and X-Forwarded-* headers
As specified in RFC 7239, parameters from the Forwarded header are
optional.
Having a Forwarded header does not imply having a forwarded host
value in it.
For this reason, even if a Forwarded header is present, it is still
needed to check for additional values provided by the X-Forwarded-*
header family, when one is missing from the Forwarded header.