CaffeineCacheManager supports async refreshing value
Prior to this commit, the current `CaffeineCacheManager` supports setting the cache builder through `set Caffeine`, but only supports the synchronous `cacheLoader`.

There is a useful feature in Caffeine: `refreshAfterWrite` which allows us to get the latest value of the cache as early as possible in the background (note that this itself is an asynchronous operation), instead of synchronously getting data from the data source when it expires.

The current `cacheLoader` can also be used to refresh the cache, but due to its synchronous blocking call characteristics, this will reduce the cache refresh rate, resulting in a lot of delay in the actual refresh time of many cache items. By using `asyncCacheLoader`, the blocking of the refresh thread can be avoided, thereby speeding up the cache refresh rate (how to implement efficient asynchronous calls is guaranteed by the `asyncCacheLoader` provider).

Because of the use of `com.github.benmanes.caffeine.cache.AsyncLoadingCache#synchronous`, it will create a new view for the cache, which will change the action of fetching the cache from asynchronous to synchronous, so this will not break the existing APIs