Make BeanNameViewResolver more extensible
With this commit we make the BeanNameViewResolver a bit more
extensible by providing an overridable method which will get
the BeanFactory/ApplicationContext to retrieve the views from.

With this extension we can now provide a JavaConfigViewResolver which
will create an AnnotationConfigWebApplicationContext for the given
configuration classes. This internally created context will be used
to retrieve the views from, with the general context registered as
a parent context.

Fixes: gh-29383