Introduce DirectShardCallbackTemplate to execute query callbacks in a shard.
This commit introduces a set of classes with the goal of providing an API to execute queries using direct shard routing. `DirectShardCallbackTemplate` provides `execute` methods that take a sharding key and a callback, so that the jdbc queries within that callback will be executed on the shard determined by the specified sharding key.