Add extension for PropertyResolver#getProperty(String, Class<T>, T)
Allows Kotlin caller to get a non-nullable property,
using a default value when the property is not set.
A method already exists on PropertyResolver which does this,
but takes a type parameter.
The extension method can eliminate the parameter via reified type,
similar to the other extension methods which exist already.

Remove @Disabled from PropertyResolverExtensionsKotlinTests
to allow tests to run