Link to KDoc API documentation from Javadoc overview
This commit also fixes the javadoc task to take the overview.html asset
into account, since this is where the link to KDoc is set up.

Note that the output directory for kdoc is changed to be consistent
with what's in the docsZip, so that the link works in local builds too.

See gh-28055